allow setting of telemetry options (#2018)
This is a proof of concept of something I'd like to be able to do with
Mastra: set telemetry settings at generation time. E.g.,

```typescript
      const res = await mastra.getAgent("cannon").stream(messages, {
          telemetryOptions: {
              functionId: "my-awesome-function", // Trace name
              metadata: {
                  langfuseTraceId: "trace-123", // Langfuse trace
                  tags: ["story", "cat"], // Custom tags
                  userId: "user-123", // Langfuse user
                  sessionId: "session-456", // Langfuse session
                  foo: "bar", // Any custom attribute recorded in metadata
              },
          },
      });
```

Effectively it's giving me more control over the telemetry options
passed to the AI SDK, which is required for traces to have these bits of
information in Langfuse
(https://langfuse.com/docs/integrations/vercel-ai-sdk#pass-custom-attributes).

Feedback welcome--happy to clean up this PR if this is a direction you
are open to!