fix(core, memory): thread title gen for models that don't support structured output (#2760)
This PR:
1. Changes thread title gen to use text output instead of
structuredOutput since not all models support that (tested it with
deepseek r1 and gpt-4o and both did well creating a title this way)
2. Adds an option to disable thread title gen since it introduces a bit
of latency `new Memory({ threads: { generateTitle: false } })`. Probably
we should flip this to default to false, left it default true for now.
This is mainly useful for building UIs around threads (like playground)
3. Adds a fireworks deepseek r1 example (shown here
https://x.com/smthomas3/status/1897044375957655695)